requirements.txt
Flask==3.0.3
keras==3.3.3
numpy==1.26.4
Pillow==10.3.0
